Move ADC code into `Digital.cpp`
Consistent with declarations in `Digital.h`
