--- Sming/Arch/Esp8266/Core/adc.cpp
@@ -1,10 +1,0 @@
-#include <Digital.h>
-#include <esp_systemapi.h>
-
-uint16_t analogRead(uint16_t pin)
-{
-	if(pin == A0)
-		return system_adc_read();
-	else
-		return -1; // Not supported
-}
